requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
boto3==1.40.20
botocore==1.40.20
celery==5.5.3
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
crispy-tailwind==1.0.3
cryptography==44.0.0
daphne==4.1.2
data-wizard==2.0.0
defusedxml==0.8.0rc2
diff-match-patch==20241021
dj-rest-auth==6.0.0
Django==4.2.19
django-admin-rangefilter==0.13.2
django-allauth==65.9.0
django-autoslug==1.9.9
django-ckeditor==6.7.2
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-filter==24.3
django-import-export==4.3.3
django-js-asset==2.2.0
django-nested-admin==4.1.1
django-redis==5.4.0
django-rest-multiple-models==2.1.3
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
et_xmlfile==2.0.0
gunicorn==23.0.0
h11==0.14.0
hiredis==3.1.0
html-json-forms==1.1.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itertable==2.2.0
jmespath==1.0.1
kombu==5.5.4
Markdown==3.7
msgpack==1.1.0
natural-keys==2.1.1
numpy==2.2.6
oauthlib==3.2.2
opencv-python==4.12.0.88
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
prompt_toolkit==3.0.52
psycopg==3.1.18
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-monkey-business==1.1.0
python3-openid==3.2.0
pytz==2024.2
redis==6.4.0
requests==2.32.5
requests-oauthlib==2.0.0
s3transfer==0.13.1
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
tablib==3.7.0
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
websockets==14.2
whitenoise==6.8.2
wsproto==1.2.0
xlrd==2.0.1
zoomus==1.2.1
zope.interface==7.2
reportlab==4.2.5
